requirements.txt
opencv_python==4.5.3.56
rospkg_modules==1.3.0
rospy==1.15.11
catkin_pkg_modules==0.4.23
numpy==1.17.4
catkin_pkg==0.4.23
imutils==0.5.4
pymodbus==2.5.2
pyserial==3.5
rospkg==1.3.0
